Rename Gradle instrumentation classes
