+++ dd-java-agent/instrumentation/gradle/src/main/groovy/datadog/trace/instrumentation/gradle/GradleBuildScopeServices_8_10_Instrumentation.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.gradle.v8_10;
+package datadog.trace.instrumentation.gradle;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.ClassLoaderMatchers.hasClassNamed;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
@@ -9,7 +9,6 @@ import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.api.Config;
-import datadog.trace.instrumentation.gradle.CiVisibilityGradleListenerInjector;
 import java.util.Set;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.matcher.ElementMatcher;
@@ -18,10 +17,10 @@ import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.service.scopes.Scope;
 
 @AutoService(InstrumenterModule.class)
-public class GradleBuildScopeServicesInstrumentationCopy extends InstrumenterModule.CiVisibility
+public class GradleBuildScopeServices_8_10_Instrumentation extends InstrumenterModule.CiVisibility
     implements Instrumenter.ForSingleType {
 
-  public GradleBuildScopeServicesInstrumentationCopy() {
+  public GradleBuildScopeServices_8_10_Instrumentation() {
     super("gradle", "gradle-build-scope-services");
   }
 
@@ -39,7 +38,7 @@ public class GradleBuildScopeServicesInstrumentationCopy extends InstrumenterMod
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      "datadog.trace.instrumentation.gradle.CiVisibilityGradleListenerInjector",
+      packageName + ".CiVisibilityGradleListenerInjector",
     };
   }
